requirements.txt
https://github.com/Lightning-AI/lightning/archive/refs/heads/release/stable.zip
click==8.1.7
einops==0.8.0
matplotlib==3.9.2
moviepy==1.0.3
numpy==2.1.3
opencv_python==4.10.0.84
Pillow==11.0.0
pytorch_lightning==2.4.0
PyYAML==6.0.2
scikit_learn==1.5.2
tabulate==0.9.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
